requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.177.0 
